backend/requirements.txt
asgiref==3.7.2
Django==4.2.6
djangorestframework==3.14.0
yfinance==0.2.31
psycopg2-binary==2.9.9
django-cors-headers==4.3.0
